Added Maven Support + org.json Added utilities packages with: JSONHandler - read and write to local json. CartObject - an object to save all of each cart items to be.
