+++ src/main/java/main/java/com/hit/IAlgoKnapsack.java
@@ -4,7 +4,7 @@ import java.util.Collection;
 public interface IAlgoKnapsack {
 
  Collection<Integer> createShoppingCart(int[] v, int[] w,
-                                              int totalWeight, int n);
+                                        int totalWeight, int n);
 }
